Run onedir pytest fixture only on module tests
